refactor: estrai domain/report.py da cli/inspect/
Sposta report_ops (build/write run report, dataset readme) in domain/report.py.
cli/inspect/report_ops.py diventa re-export puro.
Azzera dipendenza CLI interna per report_ops.

4 file, 0 logica modificata.
